move configMessage handling to a another file
